common eap code
