+++ project-integrations/aerogear-ups/app/src/main/webapp/WEB-INF/jboss-web.xml
@@ -1,5 +1,5 @@
 <jboss-web>
     <valve>
-        <class-name>org.keycloak.adapters.as7.KeycloakAuthenticatorValve</class-name>
+        <class-name>org.keycloak.adapters.jbossweb.KeycloakAuthenticatorValve</class-name>
     </valve>
 </jboss-web>
